🧪 Test: e2e login routes
