+++ server/test/helpers/createUsersAndLogin.js
@@ -1,7 +1,7 @@
 const request = require("supertest");
 const app = require("../../src/app");
 const User = require("../../src/models/userModel");
-const { testUsers } = require("../testData");
+const {testUsers} = require("../testData");
 
 const createUsersAndLogin = async (numOfUsers) => {
   const users = testUsers.slice(0, numOfUsers);
@@ -9,7 +9,7 @@ const createUsersAndLogin = async (numOfUsers) => {
 
   for (const user of users) {
     // Signup user
-    const { role, ...userData } = user;
+    const {role, ...userData} = user;
     const signupRes = await request(app)
       .post("/api/auth/signup")
       .send(userData);
@@ -32,9 +32,12 @@ const createUsersAndLogin = async (numOfUsers) => {
         .get("set-cookie")
         .find((cookie) => cookie.startsWith("accessToken="))
         .match(/accessToken=([^;]+)/)[1] || null;
-    loggedInUsers.push({ ...loginRes.body.user, accessToken });
-
-    loggedInUsers.push({ ...loginRes.body.user, accessToken });
+    loggedInUsers.push({
+      ...loginRes.body.user,
+      accessToken,
+      email: userData.email,
+      password: userData.password,
+    });
   }
 
   return loggedInUsers;
