+++ server/test/unit/controllers/authController.test.js
@@ -36,71 +36,6 @@ afterAll(async () => {
 });
 
 describe("authController", () => {
-  describe("signUp", () => {
-
-
-  describe("/login route", () => {
-    const login = async (userData) =>
-      request(app).post("/api/auth/login").send(userData);
-    let accessToken, refreshToken;
-
-    it("should login a user and return access and refresh tokens", async () => {
-      const res = await login({
-        email: "john@example.com",
-        password: "Pa$$1234",
-      });
-
-      expect(res.status).toBe(200);
-      expect(res.body.status).toBe("success");
-      expect(res.body.user).toHaveProperty("name", "John Doe");
-
-      const tokens = validateAndExtractTokens(res).validate();
-      accessToken = tokens.accessToken;
-      refreshToken = tokens.refreshToken;
-    });
-
-    it("should save the new refresh token in the database", async () => {
-      const databaseRefreshToken = await RefreshToken.findOne({
-        token: refreshToken,
-      });
-
-      expect(databaseRefreshToken).toBeTruthy();
-    });
-
-    it("should fail when email or password is missing", async () => {
-      const res = await login({
-        email: "",
-        password: "",
-      });
-
-      expect(res.status).toBe(422);
-      expect(res.body.status).toBe("fail");
-      expect(res.body.message).toMatch(/Please provide email and password/i);
-    });
-
-    it("should fail when user does not exist", async () => {
-      const res = await login({
-        email: "nonexistent@example.com",
-        password: "Pa$$1234",
-      });
-
-      expect(res.status).toBe(401);
-      expect(res.body.status).toBe("fail");
-      expect(res.body.message).toMatch(/Invalid email or password/i);
-    });
-
-    it("should fail when password is incorrect", async () => {
-      const res = await login({
-        email: "john@example.com",
-        password: "wrongpassword",
-      });
-
-      expect(res.status).toBe(401);
-      expect(res.body.status).toBe("fail");
-      expect(res.body.message).toMatch(/Invalid email or password/i);
-    });
-  });
-
   describe("/refresh-token route", () => {
     let accessToken, refreshToken;
